@@ -5,19 +5,16 @@ import androidx.room.ForeignKey
 import androidx.room.Ignore
 import androidx.room.PrimaryKey
 import com.pirateman.exercisemanager.ExerciseInterval
+import com.pirateman.exercisemanager.interval.Interval
 import java.io.Serializable
 
 @Entity
-class Exercise(var name: String?, var muscleGroup: String?, var method: String?, var selected: Boolean = false) : Serializable {
+class Exercise(var name: String = "", var muscleGroup: String = "", var method: String = "", var selected: Boolean = false) : Serializable {
     @PrimaryKey(autoGenerate = true)
-    var id = 0
-
-    @ForeignKey(entity = ExerciseInterval::class, parentColumns = ["id"], childColumns = ["rowId"])
-    @Ignore
-    private val intervalIds: List<Int>? = null
+    var id: Long = 0
 
     override fun toString(): String {
-        return "$id $name $method $muscleGroup $selected"
+        return "[$id, $name, $method, $muscleGroup, $selected]"
     }
 
 
