@@ -10,22 +10,22 @@ import androidx.recyclerview.widget.RecyclerView
 import com.bignerdranch.android.multiselector.MultiSelector
 import com.bignerdranch.android.multiselector.SwappingHolder
 import com.pirateman.exercisemanager.R
+import com.pirateman.exercisemanager.databinding.FragmentAvailableExercisesBinding
+import com.pirateman.exercisemanager.databinding.RecyclerViewExerciseItemBinding
+import org.koin.java.KoinJavaComponent.inject
 
-import org.koin.core.KoinComponent
-import org.koin.core.inject
 
+//TODO: Refactor to use viewbinding
 @RequiresApi(Build.VERSION_CODES.LOLLIPOP)
-class ExerciseHolder(val view: View) : KoinComponent, SwappingHolder(view), View.OnClickListener, View.OnLongClickListener {
+class ExerciseHolder(private val binding: RecyclerViewExerciseItemBinding, private val exerciseViewModel: ExerciseViewModel) : SwappingHolder(binding.root), View.OnClickListener, View.OnLongClickListener {
 
-    val tvName = view.findViewById(R.id.tvName) as TextView
-
-    private val exerciseViewModel: ExerciseViewModel by inject()
+    val tvName = binding.tvName
 
     init {
-        view.isClickable = true
-        view.setOnClickListener(this)
-        view.isLongClickable = true
-        view.setOnLongClickListener(this)
+        binding.root.isClickable = true
+        binding.root.setOnClickListener(this)
+        binding.root.isLongClickable = true
+        binding.root.setOnLongClickListener(this)
     }
 
     @RequiresApi(Build.VERSION_CODES.M)
@@ -36,14 +36,12 @@ class ExerciseHolder(val view: View) : KoinComponent, SwappingHolder(view), View
         val p = itemView.parent as View
 
         val rv = p.findViewById(R.id.rvExerciseList) as RecyclerView
-        rv.layoutManager = object : LinearLayoutManager(itemView.context){
-            override fun canScrollHorizontally(): Boolean {
-                return false
-            }
-
-        }
-
-
+//        rv.suppressLayout(true)
+//        rv.layoutManager = object : LinearLayoutManager(itemView.context){
+//            override fun canScrollVertically(): Boolean {
+//                return false
+//            }
+//        }
 
         val v = itemView.tag as Exercise
 
@@ -55,7 +53,7 @@ class ExerciseHolder(val view: View) : KoinComponent, SwappingHolder(view), View
 
             val color = context.resources.getColor(R.color.colorPrimary, context.theme)
             val drawable = ColorDrawable(color)
-            view.setBackgroundColor(color)
+            itemView.setBackgroundColor(color)
             selectionModeBackgroundDrawable = drawable
 
         } else {
@@ -66,7 +64,7 @@ class ExerciseHolder(val view: View) : KoinComponent, SwappingHolder(view), View
 
             val color = context.resources.getColor(R.color.white, context.theme)
             val drawable = ColorDrawable(color)
-            view.setBackgroundColor(color)
+            itemView.setBackgroundColor(color)
             selectionModeBackgroundDrawable = drawable
 
         }
